Refactored files to allow for server component to look for cookies
