+++ src/app/components/App.tsx
@@ -1,0 +1,143 @@
+"use client";
+
+import { EndpointForm } from "@/app/components/EndpointForm";
+import { DisplayData } from "@/app/components/DisplayData";
+import { useState, useEffect } from "react";
+import { SessionProvider } from "next-auth/react";
+import QueryContainer from "./QueryContainer";
+import { ChildData, ClickField, Data } from "../../../types";
+import SaveModal from "./SaveModal";
+import SaveResponseModal from "./SaveResponseModal";
+import { jsonToGraphQLQuery } from "json-to-graphql-query";
+import TutorialModal from "./TutorialModal";
+import { checkForCookies } from "@/utils/checkForCookies";
+
+export default function App({ session, cookie }: any) {
+  // data fetching: https://youtu.be/gSSsZReIFRk?t=293
+  const [data, setData] = useState<Data>({schema:{fields: [], types: {}}, endpoint:""});
+  //@ts-ignore
+  const [clickField, setClickField] = useState<ClickField>({ type: "", field: "" });
+  const [isSaveModalOpen, setIsSaveModalOpen] = useState(false);
+  const [isSaveResponseModalOpen, setIsSaveResponseModalOpen] = useState(false);
+  const [queryAsString, setQueryAsString] = useState("query: { \n \n }");
+  const [queryAsObj, setQueryAsObj] = useState({ query: {} });
+  const [argument, setArgument] = useState({});
+  const [saveResponseStatus, setSaveResponseStatus] = useState(false);
+  const [isTutorialModalOpen, setIsTutorialModalOpen] = useState(cookie);
+
+  const childToParent = (childData: ChildData): void => {
+    setData(childData);
+  };
+  
+  useEffect(() => {
+    if (Object.keys(argument).length > 0) {
+      console.log('argument changed', argument);
+      console.log('query as object', queryAsObj);
+      const queryAsObjDeepCopy = JSON.parse(JSON.stringify(queryAsObj));
+      console.log('queryAsObjDeepCopy',queryAsObjDeepCopy);
+      //@ts-ignore
+      queryAsObjDeepCopy.query[argument.field]["__args"][argument.argument] = argument.value;
+      setQueryAsObj(queryAsObjDeepCopy);
+    }
+  },[argument]);
+
+  useEffect(() => {
+    setQueryAsString(jsonToGraphQLQuery(queryAsObj, { pretty: true }));
+  }, [queryAsObj]);
+
+  useEffect(() => {
+    // if the clicked node should update the query generator...
+    if(clickField.field.length > 0){
+      console.log('clickField', clickField);
+      if(clickField.field && (clickField.type || clickField.data.queryField)) {
+        //if a field of the query type is clicked, should check for arguments
+        if (clickField.data.queryField) {
+          updateQueryAsObj("", clickField.field.toLowerCase(), clickField.data.arguments);
+        }
+        //if clicked node is an argument, shouldn't update query generator
+        else if (!clickField.data.isArg) {
+          // console.log('clicked is not an argument')
+          updateQueryAsObj(clickField.field.toLowerCase(), clickField.type.toLowerCase(), []);
+        }
+      }
+    }
+  }, [clickField]);
+
+  const updateQueryAsObj = (child: string, parent: string, args: string[]) => {
+    //make a deep copy of queryAsObj
+    const tempObj = JSON.parse(JSON.stringify(queryAsObj));
+    console.log('tempObj1', tempObj);
+    //@ts-ignore
+    if (!tempObj.query[parent]) {
+      tempObj.query[parent] = {};
+    }
+    if(child !== ""){
+      tempObj.query[parent][child] = true;
+    }
+    
+    // add "__args" object if args are required
+    if (args) {
+      tempObj.query[parent]["__args"] = {};
+      for(let arg of args) {
+        tempObj.query[parent]["__args"][arg] = "";
+      }
+    }
+    
+    // console.log('tempObj2', tempObj);
+      
+    //update object in state
+    setQueryAsObj(tempObj);
+    //update string in state
+    setQueryAsString(jsonToGraphQLQuery(tempObj, { pretty: true }));
+  };
+
+  /*
+  CHECK IF THE USER HAS VISITED BEFORE VIA THE COOKIE 
+  */
+
+  console.log(cookie);
+
+  return (
+    <>
+      <SessionProvider session={session}>
+        {!isTutorialModalOpen && <TutorialModal 
+        isTutorialModalOpen={isTutorialModalOpen}
+        setIsTutorialModalOpen={setIsTutorialModalOpen}
+        />}
+        <EndpointForm childToParent={childToParent} />
+        <div className="grid grid-cols-3">
+          <div className="col-span-2 dark:bg-slate-800">
+            {
+              data.schema.fields.length &&
+              <DisplayData
+                data={data}
+                //@ts-ignore
+                setClickField={setClickField}
+                setData={setData}
+                queryAsObj={queryAsObj}
+                //@ts-ignore
+                setQueryAsObj={setQueryAsObj}
+                setArgument={setArgument}
+                // argModified={argModified}
+              />
+            }
+          </div>
+          <div className="h-screen dark:bg-slate-800">
+            <QueryContainer
+              endpoint={data.endpoint}
+              clickField={clickField}
+              setIsSaveModalOpen={setIsSaveModalOpen}
+              queryAsString={queryAsString}
+              setQueryAsString={setQueryAsString}
+              //@ts-ignore
+              queryAsObj={queryAsObj}
+              setQueryAsObj={setQueryAsObj}
+            />
+          </div>
+        </div>
+        <SaveModal isSaveModalOpen={isSaveModalOpen} setIsSaveModalOpen={setIsSaveModalOpen} setIsSaveResponseModalOpen={setIsSaveResponseModalOpen} setSaveResponseStatus={setSaveResponseStatus} query={queryAsString} endpoint={data.endpoint} />
+        <SaveResponseModal isSaveResponseModalOpen={isSaveResponseModalOpen} setIsSaveResponseModalOpen={setIsSaveResponseModalOpen} saveResponseStatus={saveResponseStatus}/>
+      </SessionProvider>
+    </>
+  );
+}
